requirements_SHBHHBSH.txt
numpy>=1.21.0
scipy>=1.7.0
pandas>=1.3.0
matplotlib>=3.4.0
pyaudio>=0.2.11
librosa>=0.8.1
sounddevice>=0.4.1
rtlsdr>=0.2.93
python-nmap>=0.7.1
scapy>=2.4.5
psutil>=5.8.0
GPUtil>=1.4.0
pyserial>=3.5
opencv-python>=4.5.0
folium>=0.12.0
geopy>=2.2.0
requests>=2.25.0
python-dateutil>=2.8.0
Pillow>=8.3.0
